Improve Move exceptions and exception utility classes to the streampark-console
