+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/base/exception/AbstractApiException.java
@@ -15,16 +15,35 @@
  * limitations under the License.
  */
 
-package org.apache.streampark.common.exception;
+package org.apache.streampark.console.base.exception;
 
-/** System Internal Exception */
-public class InternalException extends Exception {
+/**
+ *
+ *
+ * <pre>
+ * An exception message that needs to be notified to front-end.
+ * </pre>
+ */
+public abstract class AbstractApiException extends RuntimeException {
 
-  public InternalException(String message) {
+  private final long responseCode;
+
+  protected AbstractApiException(String message, long responseCode) {
     super(message);
+    this.responseCode = responseCode;
   }
 
-  public InternalException(String message, Throwable cause) {
+  protected AbstractApiException(Throwable cause, long responseCode) {
+    super(cause);
+    this.responseCode = responseCode;
+  }
+
+  protected AbstractApiException(String message, Throwable cause, long responseCode) {
     super(message, cause);
+    this.responseCode = responseCode;
+  }
+
+  public long getResponseCode() {
+    return responseCode;
   }
 }
