+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/base/exception/ApiAlertException.java
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.streampark.common.exception;
+package org.apache.streampark.console.base.exception;
 
-import org.apache.streampark.common.domain.ResponseCode;
+import org.apache.streampark.console.base.domain.ResponseCode;
 
 import java.util.Objects;
