+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/controller/FlinkClusterController.java
@@ -18,8 +18,8 @@
 package org.apache.streampark.console.core.controller;
 
 import org.apache.streampark.common.enums.ClusterState;
-import org.apache.streampark.common.exception.InternalException;
 import org.apache.streampark.console.base.domain.RestResponse;
+import org.apache.streampark.console.base.exception.InternalException;
 import org.apache.streampark.console.core.bean.ResponseResult;
 import org.apache.streampark.console.core.entity.FlinkCluster;
 import org.apache.streampark.console.core.service.FlinkClusterService;
