+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/entity/FlinkEnv.java
@@ -18,9 +18,9 @@
 package org.apache.streampark.console.core.entity;
 
 import org.apache.streampark.common.conf.FlinkVersion;
-import org.apache.streampark.common.exception.ApiDetailException;
 import org.apache.streampark.common.util.DeflaterUtils;
 import org.apache.streampark.common.util.PropertiesUtils;
+import org.apache.streampark.console.base.exception.ApiDetailException;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
