+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/service/alert/impl/EmailAlertNotifyServiceImpl.java
@@ -17,8 +17,7 @@
 
 package org.apache.streampark.console.core.service.alert.impl;
 
-import org.apache.streampark.common.exception.AlertException;
-import org.apache.streampark.common.util.PremisesUtils;
+import org.apache.streampark.console.base.exception.AlertException;
 import org.apache.streampark.console.base.util.FreemarkerUtils;
 import org.apache.streampark.console.base.util.PremisesUtils;
 import org.apache.streampark.console.core.bean.AlertConfigParams;
