+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/system/service/impl/UserServiceImpl.java
@@ -17,13 +17,12 @@
 
 package org.apache.streampark.console.system.service.impl;
 
-import org.apache.streampark.common.domain.ResponseCode;
-import org.apache.streampark.common.exception.ApiAlertException;
 import org.apache.streampark.common.util.DateUtils;
-import org.apache.streampark.common.util.PremisesUtils;
 import org.apache.streampark.common.util.Utils;
+import org.apache.streampark.console.base.domain.ResponseCode;
 import org.apache.streampark.console.base.domain.RestRequest;
 import org.apache.streampark.console.base.domain.RestResponse;
+import org.apache.streampark.console.base.exception.ApiAlertException;
 import org.apache.streampark.console.base.mybatis.pager.MybatisPager;
 import org.apache.streampark.console.base.properties.ShiroProperties;
 import org.apache.streampark.console.base.util.PremisesUtils;
